tests: dns_resolve: Make testcase looks similar to other testcases
All test case should have entry function in main.c file and
other functions should be called from separate file.

It also removes return statement in void function.

JIRA: ZEP-2249
