+++ tests/net/lib/dns_resolve/src/Makefile
@@ -4,4 +4,4 @@ ccflags-y += -I${ZEPHYR_BASE}/tests/include
 
 include $(ZEPHYR_BASE)/tests/Makefile.test
 
-obj-y = main.o
+obj-y = main.o test_dns_resolve.o
